Allow setting MYSQL_PASSWORD via file contents.
This is analagous to what is done with `MYSQL_ROOT_PASSWORD`. If
`MYSQL_PASSWORD` is an existing file path, then use the contents
of the file as the value for the password when creating a
non-root user.

This makes it possible to use Docker secrets as a way to pass
this particular credential into the `mysql-server` container.