Fix bug#37821740
When running mysqld --validate-config we are not passing the --user
option. Thus the server has no chance to run as target user, thus
if any component writes to filesystem this may be done as wrong user.

However server doesn't fully recognize this flag. For this working
properly we also need a fix to 37921092.

Change-Id: Ib30bfbef85abbaa1c4900c14b9ad7b734a44bb69